[REFACT] facade 패턴 적용
