+++ src/main/java/com/programmers/smrtstore/domain/user/application/UserFacade.java
@@ -1,0 +1,20 @@
+package com.programmers.smrtstore.domain.user.application;
+
+import com.programmers.smrtstore.domain.user.presentation.dto.req.DetailShippingRequest;
+import com.programmers.smrtstore.domain.user.presentation.dto.req.UpdateUserRequest;
+import com.programmers.smrtstore.domain.user.presentation.dto.res.DeliveryAddressBook;
+import com.programmers.smrtstore.domain.user.presentation.dto.res.DetailShippingResponse;
+import com.programmers.smrtstore.domain.user.presentation.dto.res.ProfileUserResponse;
+
+public interface UserFacade {
+
+    ProfileUserResponse getUserInfo(Long userId);
+    ProfileUserResponse update(Long userId, UpdateUserRequest request);
+    void withdraw(Long userId);
+
+    DetailShippingResponse createShippingAddress(Long userId, DetailShippingRequest request);
+    DeliveryAddressBook getShippingAddressList(Long userId);
+    DetailShippingResponse updateShippingAddress(Long userId, Long shippingId, DetailShippingRequest request);
+    DetailShippingResponse findByShippingId(Long shippingId);
+    void deleteShippingAddress(Long userId, Long shippingId);
+}
