+++ src/main/java/com/programmers/smrtstore/domain/user/presentation/UserController.java
@@ -1,6 +1,7 @@
 package com.programmers.smrtstore.domain.user.presentation;
 
 import com.programmers.smrtstore.common.annotation.UserId;
+import com.programmers.smrtstore.domain.user.application.UserFacade;
 import com.programmers.smrtstore.domain.user.application.UserService;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.DetailShippingRequest;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.UpdateUserRequest;
@@ -23,12 +24,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final UserFacade userFacade;
 
     @GetMapping("/profile")
     public ResponseEntity<ProfileUserResponse> profile(
         @UserId Long userId) {
-        ProfileUserResponse response = userService.getUserInfo(userId);
+        ProfileUserResponse response = userFacade.getUserInfo(userId);
         return ResponseEntity.ok(response);
     }
 
@@ -36,13 +37,13 @@ public class UserController {
     public ResponseEntity<ProfileUserResponse> update(
         @UserId Long userId, @RequestBody @Valid
     UpdateUserRequest request) {
-        ProfileUserResponse response = userService.update(userId, request);
+        ProfileUserResponse response = userFacade.update(userId, request);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> withdraw(@UserId Long userId) {
-        userService.withdraw(userId);
+        userFacade.withdraw(userId);
         return ResponseEntity.noContent().build();
     }
 
@@ -50,35 +51,35 @@ public class UserController {
     public ResponseEntity<DetailShippingResponse> createShippingAddress(
         @UserId Long userId,
         @RequestBody @Valid DetailShippingRequest request) {
-        DetailShippingResponse response = userService.createShippingAddress(userId, request);
+        DetailShippingResponse response = userFacade.createShippingAddress(userId, request);
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("/shipping")
     public ResponseEntity<DeliveryAddressBook> shippingAddressList(
         @UserId Long userId) {
-        DeliveryAddressBook response = userService.getShippingAddressList(userId);
+        DeliveryAddressBook response = userFacade.getShippingAddressList(userId);
         return ResponseEntity.ok(response);
     }
 
     @PostMapping("shipping/{shippingId}")
     public ResponseEntity<DetailShippingResponse> updateDefaultShippingAddress(@UserId Long userId,
         @PathVariable Long shippingId, @RequestBody @Valid DetailShippingRequest request) {
-        DetailShippingResponse response = userService.updateShippingAddress(userId, shippingId,
+        DetailShippingResponse response = userFacade.updateShippingAddress(userId, shippingId,
             request);
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("/shipping/{shippingId}")
     public ResponseEntity<DetailShippingResponse> findByShippingId(@PathVariable Long shippingId) {
-        DetailShippingResponse response = userService.findByShippingId(shippingId);
+        DetailShippingResponse response = userFacade.findByShippingId(shippingId);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/shipping/{shippingId}")
     public ResponseEntity<Void> deleteShippingAddress(@UserId Long userId,
         @PathVariable Long shippingId) {
-        userService.deleteShippingAddress(userId, shippingId);
+        userFacade.deleteShippingAddress(userId, shippingId);
         return ResponseEntity.noContent().build();
     }
 }
